requirements.txt
aws-cdk-lib==2.45
aws-cdk.aws-appsync-alpha==2.45.0a0
python-dotenv==0.21.0
constructs==10.1.129
boto3==1.24.89
botocore==1.27.89
black==22.10.0
